Fix the broken category in certain circumstances
This occurs when the the UI language differs from the wikis
content language. The code uses the namespace name for the
category namespace based on the UI language. When a user has
a different language set in their preferences than the wiki
language, the category added will be in the wrong language
resulting in a broken link and uncategorized video.

This patch also introduces short array syntax.

Bug: T179332
Change-Id: I5e87e92e09b3c90e890e76e66066085d3006556e