Update our forked version of Special:Undelete from MW 1.24.1 version to 1.30.0 version
The fact that we even *have* to copy this much code just to add support
for a new namespace to Special:Undelete is absurd and ridiculous.

Mostly copypasted from core, the changes specific to the Video extension
are appropriately marked as such with the "CORE HACK" tags to easily
identify them.

Untested.

Change-Id: Ic6b7b494ec8bbd416d0488ed52f54dd49348c241