Fix video undeletion support for MW 1.31+
The main reason as to why this class exists is so that we can use our custom VideoPageArchive class instead of core PageArchive in a few places.
That class is initialized four times in total in four different methods, all of which do quite a bit of processing.
Because of that as well as the fact that many of these methods are currently private, we need to copy 'em over and that is hardly ideal.

Change-Id: I91fac869c09e3dcbea330b14f138b94e9c012431