Fix Special:Undelete...by copying over the entire SpecialUndelete.php from REL1_32 and applying our custom hacks on top of it
This is really awful in general and makes for a messy diff, but it appears to be the only foolproof way for sure.

Prior to this change, the entire Special:Undelete page was broken, since
1. we make the special page use this class (SpecialUndeleteWithVideoSupport) instead of the core SpecialUndelete,
2. but formatFileRow() is private in SpecialUndelete,
3. and thus we can't use it from here.

Even if that particular function would be protected and thus overrideable from here, there'd still be at least one bug: for deleted (image) files, the URL would not show up, thus making you unable to view deleted images as an admin with sufficient privileges.

I'm guessing that there are more pitfalls and whatnot in here and it seems that this is the "easiest" way around them for now: copy core code upon a new major MW release, apply hacks on top of it and call it a day until a new major version is released.

Change-Id: I4909e4a951a98a4952eb194e320468e1918b7ae0