Use preferred methods to HTML escape messages
I believe it was possible to inject arbitrary HTML in a few of these
places by someone who manages to edit these messages.

This is how messages should be HTML escaped, to the best of my
knowledge, to make such attacks impossible.

Note I was not able to manually test all situations where these
messages are shown. I hope  someone more active in this codebase is
able to do this.

Change-Id: Ie1a9610c645f16fdeb99953d75e1a406c56c1821