Fix a few small but confusing PHPCS code style violations
In one place a Phan exclusion conflicts with PHPCS. I found a
"previous-line" tag that looks like it can help here.

Forcing formatRow to return null is not a good idea. The value is
used and appended to a string. The correct workaround is to return
an empty string. The effect is the same, it just avoids fragile
type juggling.

Change-Id: Ia12606fd720a558012e6721b37e5dc018ea75f07